requirements.txt
fastapi
uvicorn
websocket-client
requests